libMirage: Disc: added 'dvd-report-css' boolean property. When enabled, the fake Disc Structure 0x01 indicates that the DVD disc is CSS-encrypted.
Obviously, this works only with DVD discs and only with images that don't provide disc structure data.

This is the first step towards enabling playback from 'raw' images of CSS-encrypted DVDs - the ones created using 'dd' and without CSS stripped. It turns out that in order for such images to be played from a CDEmu device using e.g. mplayer2, we don't need the whole key exchange infrastructure - it's enough for us to signal that the image is encrypted and then libdvdcss will do brute-force decryption of it :)