MacBinary file-filter: Researched and implemented routines for parsing NDIF data structures.
NDIF is the (old) format for Apple Disk images, used by .SMI and .IMG files.
The main difference from the UDIF format is how the resource-fork is stored.
UDIF stores the resource-fork internally, whereas NDIF have the resource-fork
stored externally, either in a HFS filesystem or preserved in a MacBinary etc.
The resource-fork holds important metadata about the disk image itself.

This commit adds the functionality to completely parse NDIF images.
Effectively interpreting the images will be difficult however.
According to rumors Apple uses a proprietary compression algorithm known as
KenCode, and if that is true it will make things difficult.