libmirage: B6T-parser: Add a complete set of media types.
It seems to me as long as the precise media type is documented in the form
of an enum and ideally a comment, we don't need to match each and everyone
in the media detection routine. We just need to collectively tie a bunch
of media types to their respective Mirage media type.