The Great Purge: use custom stream implementation to replace both use of GInputStream/GSeekable throughout libMirage and our old GFilterInputStream-based architecture of file filters.
So, I was looking for a way to implement a write version of our libsndfile-based file filter, using GFilterOutputStream similarly to how we use GFilterInputStream to implement read-only filter streams. The initial idea was to have an image writer to open an output filter stream chain, which creates a file, and then also open an input filter stream chain on now-created file for read-only access - by setting both chains to a fragment, we would obtain a read/write fragment (note: read access in addition to write one is desired because after CDEmu device writes an image in recording mode, it should be able to immediately read it back after the session is closed). The problem is, if output and input filter are to be separate, it is impossible to initialize the input (in this case, sndfile) filter on empty (but existing!) data (in this case, audio) file. The obvious solution, thus, would be to have a single filter object with shared state - and this is what sndfile expects us to do, too. However, although GLib provides a GIOStream object, and GFileIOStream inheriting from it, it has no such thing as an GFilterIOStream. From what I could gather, we would have to obtain separate GInputStream and GOutputStream objects from the base GIOStream, and construct separate GFilterInputStream and GFilterOutputStream chains on top of those, which makes a single-filter-object implementation rather round-about.

Therefore, time for a change. From now on, we have a MirageStream interface, which provides basic I/O methods, such as read, write, seek, and tell. In addition, it also provides get_filename method. We have two basic object types implementing this interface - MirageFileStream andMirageFilterStream. The former actually uses GFileInputStream (if read-only) or GFileIOStream (if read/write) to access files. The latter is used to implement filters (which are now renamed from "file filters" to "filter streams"), which can be used to construct filter stream chains, just as before. Therefore, former "file filters" became "filter streams".

There are several benefits:
- a filter stream can now be either read-only, or read/write, depending on what functions it implements (and on the filters below it)
- no more explicit casting to G_INPUT_STREAM and G_SEEKABLE, no more passing NULL for cancellable argument, which we did not use
- both MirageFileStream and MirageFilterStream can now inherit from MirageObject, so they do not have to implement MirageContextual (we could actually merge its functionality into MirageObject now, but at the moment, I decided to let that interface stay)
- since we now store stream's filename in MirageFileStream and propagate it up the filter stream chain, we do not have to keep stream->filename map in context object anymore (we still cache streams for performance reasons, though!)

As far as functionality of this new stream interface goes, it remains unchanged from what we had before. The only exception is the fact that we cannot chain GLib's GFilter*Stream classes on top of ours; this does not really hurt us, except in case of text-based image descriptor files, where we put GConverterInputStream and GDataInputStream on top of a GInputStream. To preserve this functionality, MirageStream interface provides mirage_stream_get_g_input_stream() method, which wraps the stream object into a "compatibilty input stream" object that inherits from GInputStream and provides a read function that just calls MirageStream's read. This way, we can keep the code in text-based image parsers more or less intact.