libMirage: TOC writer: fixed object (session, track, fragment) leak
This should in turn also fix the issue where the output streams were
not properly cleared from context's cache (due to not being freed
in the fragments). This, in turn, caused problems in scenarios where
a user would create an empty image and burn it, then remove the
resulting files, and create/burn the image again; the cache would
return old streams, which pointed to non-existant files, causing no
data to be written.