CIF-Parser: Fix track length calculation.
A test image I made showed that the allocated track length in the
image will never be less than 300 sectors (4s), this goes for both
binary and audio tracks. I also beleive that this explains all the
cases where the reported number of sectors in a track differs from
what is obviously allocated in the image. As a result this
simplifies how track length is calculated.