gCDEmu: do not try to fall back to old AppIndicator if loading AppIndicator3 fails
AppIndicator3 is based on Gtk3, while AppIndicator uses Gtk2. As
we explicitly load Gtk3 in gCDEmu now, loading AppIndicator will
cause an error in cases when AppIndicator is available, but
AppIndicator3 is not. Also, catch all errors, because
gi.require_version() throws as well.

Thanks to Nikita Zlobin for raising this issue.