libMirage: Context: expand the mirage_context_set_password_function() to include a GDestroyNotifier for user_data
Needed to make the password function callback work within the
python bindings. (The "call" annotation was definitely incorrect,
and the "async" one won't do much good, either - seems "notified"
is the way to go, and for that, we need the destroy notifier).

Also, properly initialize the password function, data, and destroy
notifier to NULL, and optionally destroy the data when the context
is destroyed.