Image analyzer: ported from C to python+GIR
The port was done for several reasons; one is to show-case the use
of libMirage GIR bindings. On the other hand, it also provides a
good test for the GIR bindings, and helps uncover incorrect
annotations in libMirage. Also, we can now use Matplotlib for disc
topology graph, and bid farewell to GnuPlot and its flaky integration
into UI via GtkSocket (which does not appear to work anymore).

Plus, the whole UI should be translatable now, if anyone feels like
it should be translated...