libMirage: Parser: removed a spurious stream reference increment when creating a text stream
At the beginning of mirage_parser_create_text_stream(), we incremented
the reference count on the input stream, but never released it.
Incrementing the count at that point should be redundant, as we take
the reference again when constructing the MirageStream->GInputStream
bridge, and then passing that to the final GDataInputStream.

Should fix Issue #97 (cdemu-daemon does not release some files),
where an application using libMirage would keep the text-based
descriptor file (CUE, CCD, TOC, etc.) open after parsing it.