libmirage: Try to prevent false positives for mime types.
The shared mime info spec IMO have multiple issues.

- There is no way to match magic bytes at a certain offset relative to
  the end of a file.
- Which match takes precedence is determined by setting a simple priority
  value that defaults to 50, and which scale is largely arbitrary.

It would be easy to improve the latter by taking into account the strength
of the match, adjusting for the probability of matching random data.
The probability (P) for a false positive when matching N bits can be given
as P = 1 / (2^N) which means you have with three bytes a one in 1.6 million
chance of a false positive, matching two or four bytes are much weaker or
stronger by far.

Which priority value should be used for matching magic strings should
as it stands assume that a typical filename suffix match (with a default
priority of 50) matches against at least 3 bytes that ideally should be
unique, and use a higher priority if the magic string is larger than that
and vice versa. And that of course ignores that some suffixes like ".img"
can mean pretty much anything and such a match is pretty much equal to
strength zero.