CMake: FindIntlTool.cmake: minor cleanup in intltool_merge() function
First, we should not be making any assumptions about location of
the given po dir - let the caller pass us the fullpath instead
(which is also consistent with intltool_process_po_files() function).

Second, make no assumption about location of the input file; i.e.,
do not try to prefix it, but rather let the caller pass us the
fullpath instead. The main issue here was that file given to the
command was prefixed with PROJECT_SOURCE_DIR, but the file specified
as the dependency was not.

Third, try to handle the cases when output file has a path component;
to generate the target name, which cannot contain slashes, we extract
the name component only.