Python programs: use gettext.install() and turn on unicode strings on python 2.7
On python 2.7 unicode strings need to be explicitly enabled with
unicode=True argument to gettext.install(); in python 3, this is
not necessary anymore, but the argument has been removed.

Up until now we did not encounter issues with this in cdemu-client
or gcdemu, but I did hit the unicode error with localized version
of image-analyzer. So to be safe and consistent, apply this change
to all our python programs.