libMirage: extend backwards-compatibility to GLib 2.28
In text-based descriptor parsers, we read text lines using
g_data_input_stream_read_line_utf8(), which was introduced by 2.30.

So on older versions, use g_data_input_stream_read_line() instead,
which does not ensure that the text is valid UTF-8. If the lack
of this check proves problematic, we can implement it ourselves -
but we will cross that bridge when we come to it (if at all).