cmake: Support installing for bash-completion v2
Adjust the install code to conform to the second version of
bash-completion. This includes:

1. using pkg-config to obtain appropriate install directory,

2. renaming the installed file to 'cdemu', in order to match
the executable name.

As a result, if bash-completion v2 is installed on user's system,
the completion file will be installed in the modern directory
and will support autoloading, in contrast to installing it to
the 'compat' bash_completion.d location that requires unconditionally
loading all completions.

To preserve compatibility, if bash-completion is not found via
pkg-config the old install location is used. Furthermore,
a CMAKE_INSTALL_COMPLETIONSDIR cache variable is introduced to allow
user to adjust the install location.