Daemon: implemented rudimentary support for BD-ROMs
Implemented Profile 0x40 (BD-ROM) and Feature 0x40 (BD Read Feature).
As per MMC-5, we should also extend the READ DISC STRUCTURE with
support for formats 0, 0xFF, and 0x30; but at the moment, this
does not seem to be critical.

This implementation is sufficient for GNOME's Disk Tool to recognize
our virtual device as "CD/DVD/Blu-Ray drive", and to recognize the
loaded Blu-Ray image as BD-ROM disc.