libMirage: ISO parser: improve sector type detection
Fixed the ISO9660/UDF descriptor pattern search for sector sizes
of 2332 and 2336 (these contain Mode2 subheader, so we need to
skip 8 bytes to get to user data).

Fixed the descriptor pattern search for sector size of 2352 and
Mode 2 sectors. Contrary to Mode 1 sectors, where we need to
skip 16 bytes (sync+header), we need to skip additional 8 bytes
(sub-header). So we perform two tests, one with 16-byte and one
with 24-byte header.

TODO: I suspect this parametrization also fully determines the
track type, so we may altogether forego the check with
mirage_parser_iso_determine_track_type().