libMirage, Daemon: disable -Wcast-function-type
With gcc 8.0, we get a lot of warnings about incompatible function
type cast because GClassInitFunc and GInstanceInitFunc take
additional parameter that is usually unused, and hence omitted.
These could be fixed by either adding those unused arguments, or
via an intermediate cast to a wrapper with the right number of
arguments.

Furthermore, we get a warning when using g_source_set_callback()
in the daemon, which cannot be fixed, because the callback
signature depends on the actual source type.

So for now, disable the warning altogether.