Daemon: re-implemented disc structure fabrication from libMirage
Also reorganized the functions a bit so that all structure
generation is handled in one place, and READ DISC STRUCTURE
command implementation is concerned only with retrieveing data.

TODO: The DVD Copyright Info structure depends on a user-provided
setting of whether the DVD should report CSS or not. This was
previously passed to the MirageDisc as a parser parameter. Now,
we provide placeholder setting in the CDemuDevice. This setting
needs to be wired in and previous approach needs to be removed
from the clients!