libMirage: MDS parser: cleanup parsing of disc structures
Do not rely on structures from disc-structures.h header which
will be removed. In practice, the structure interpretation is
needed only to read the number of layers from the structure 0x0000.

This cleanup should also fix a mismatch between structure data
and declared size when the structures were being set to the disc...