VHBA: removed the use of scsi_cmd->serial_number
Kernel 5.1 removed this field. Even so, the documentation in older
kernels for the field indicates that "It is a bug for LLDDs to use
this number for purposes other than printk (and even that is only
useful for debugging).", so we definitely should not be using it
to internally tag our commands in the queue.

Instead, each device now has a private command counter, which is
used to tag the commands in the queue (and the outgoing requests
and incoming responses in the kernel<->userspace communication).