cdemu-client: improved setting and retrieval of daemon/library debug masks
In addition to specifying the new debug mask as a single numeric
value (0x03), we can now also specify it as a string of pipe-separated
values ("0x01|0x02") or identifiers (e.g., "DAEMON_DEBUG_DEVICE|DAEMON_DEBUG_MMC").
This makes it easier to set a composite of debug masks without
having to compute the combined value.

Similarly, when displaying debug masks, we now print both the
combined value, and pipe-separated list of identifiers.