vhba: Deprecate support for Linux kernels older than 3.16
The Linux kernel 2.6.x series is between 8-16 years old at this point so this
seems like a good time to drop support for it. Upstream support for the 2.6.32
release ended March 2016, so the series as a whole is vulnerable to
Spectre and Meltdown as well since those became public knowledge January 2018.

The 3.x series of kernels is about 4-8 years old at this point and the 3.16
release is actually still supported upstream until April 2020. There was only
one small workaround for kernels older than 3.15, and normally I would have
encouraged keeping support for the 3.x series for several more years,
but considering all kernels older than 3.16 as far as I'm aware is vulnerable
to Spectre and Meltdown, encourages dropping support for any older kernels.