cdemu-client: Improve password handling in the client.
The way passwords were handled before had a number of issues. This patch
adresses the most major ones, but avoids overdoing it.

Previously one could supply a password via a command line parameter. The
command line shell keeps a history of all commands entered, and naturally
the password will show up there in plain text. On many systems the user's
home directory is readable to others by default and so is the command line
history.

The password will also show up on the 'ps' command. On many systems the
'ps' command is also not considered privileged.

This patch removes the plaintext 'password' argument, so that you will
always be prompted for a password when you try to load an encrypted image.
In case the client is not started from a terminal, the user will be prompted
from a GUI instead. The program responsible for this is specified with
the environment variable SSH_ASKPASS which is usually already configured
on most systems. For non-interactive uses it is possible to override this
with the CDEMU_ASKPASS variable (consult the README).