vhba: Handle command queue locking ourselves.
Prior to version 2.6.37 of the Linux kernel, the queuecommand method
for receiving SCSI commands from the upper SCSI layers was protected by
a spinlock allowing only one command to be passed at a time.
This spinlock was then removed to allow HBA drivers to do their own locking.

At the time this caused a crash in our driver that was fixed by essentially
adding some code that effectively put back this spinlock.

Upon closer examination it would appear that the two existing command
spinlocks are sufficient to protect threads submitting commands from stepping
on each others toes.

The original bugreport mentioned null pointer dereference causing a segfault,
and this happened immediately after loading the daemon. This is suspected to
have happened because the old and the new queuecommand methods have different
function parameters that were supplied to them via a function pointer.
Assuming the compiler allowed this to be compiled it would certainly explain
the crash.