vhba: Change how command matching and tagging is performed.
The blk-mq patchset that was added with kernel 3.17 included a facility for
assigning queue tags that are per-host. While this might seem like a strange
thing to do, Christoph Hellwig explains in the commit message: "Due the
preallocated requests the MQ code path exclusively uses the host-wide shared
tag allocator instead of a per-LUN one."

So we start using this to generate per-host metatags when it is available.
For compatibility with older kernels we adopt a similar scheme for numbering
metatags, the index number of the command in the queue.

Note that these metatags, as they now are named, is only meant for matching
a response to a request, similar to what I_T_L_Q tables are used for on real
hardware, they should not be confused with queue tags which are per-lun tags.