libMirage: MirageWriter: disambiguate open_image() method
The mirage_writer_open_image() method overlaps with the virtual
open_image() method provided by the children classes, and the
arguments between the two differ - the virtual method does not
take parameters hash table, because that's handled in the
mirage_writer_open_image(). This may confuse introspection-generated
bindings, so we should disambiguate the two, by renaming the virtual
method to open_image_impl().

Fixes the warning about conflicting method with the same name when
building VAPI file for Vala bindings.