libMirage: READCD parser: clear prev_mode when opening new session
Clear/reset the `prev_mode` variable when opening new session, in order
to prevent a spurious mode-change check on the first track of the new
session. This in turn fixes segmentation fault when the mode changes
between the last track of the previous session and the first track of
the new session, and the mode-change handler tries to operate on the
track that comes before the first track in the new session, resulting
in NULL pointer dereference.

For the sake of consistency, also clear the `leadout_lba` variable.