libMirage: ISO writer: fix segfault when track has no associated filename
Make a copy of the "<ERROR>" string that the writer uses as a
fall-back when trying to construct the array of disc filenames and a
track has no fragment with associated filename.

The `filenames` array is later free'd using `g_strfreev`, so it
needs to contain a free-able copy of the string, otherwise it
segfaults.