libMirage: writer: adjust detection of pregap-only fragments
Change the criterion for detection of pregap(-only) fragments; these
should be fragments that lie completely before the track start. The old
criterion was for fragment start to lie before the track start; but this
fails to account for possibility of having a fragment that covers both
the pregap and the data part (and in such cases, mis-classification as
pure-pregap fragment would result in data for the track not being
written to a file at all).