daemon: try using sleep inhibitor lock
Try to obtain sleep inhibitor lock from `org.freedesktop.login1.Manager`
interface at `/org/freedesktop/login1`, and use it in conjuction with
the "prepare-for-sleep" handler, as outlined in document at [1].

[1] https://systemd.io/INHIBITOR_LOCKS

I.e., when we connect to the `Manager` proxy during daemon
initialization, connect the  "prepare-for-sleep" signal, and also
try obtaining the delay variant of the sleep inhibitor lock. In
the signal handler, when entering sleep, stop the devices, and
release the inhibitor lock. When exiting sleep, try to re-acquire
the inhibitor lock, and then re-start the devices. This way, we
attempt to ensure that the system does not enter sleep before the
daemon has managed to stop all devices.

The inhibitor locks are realized as POSIX file descriptors, and
support for their transfer in GDBus bindings requires GLib >= 2.64
(and we need to pass `--glib-min-required 2.64` to `gdbus-codegen`
to generate compatible bindings).

This should not be a problem on contemporary linux distributions
anyway (Debian bullseye/oldoldstable has 2.66), but once this
suspend handler feature is finalized, we can put it under a build-time
option and allow users to disable it.