daemon: include DAEMON_DEBUG_SLEEP_HANDLER in EnumDaemonDebugMasks
Include the newly added `DAEMON_DEBUG_SLEEP_HANDLER` debug mask
value in list that we send in response to `EnumDaemonDebugMasks`
command.

Primarily for the sake of consistency and possible future extensions;
at the moment, this mask is applied only on daemon's debug context,
whereas the output of EnumDaemonDebugMasks is typically used to
control debug mask on the devices' debug contexts.