daemon: allow the new system-sleep handler to be disabled at build time
Add new CMake build option (`ENABLE_LOGIND_SLEEP_HANDLER`) that
allows the daemon to be built without support for the newly-added
system-sleep handler.

This allows daemon to be still built for systems with GLib older than
2.64 (which is required by the sleep handler), as before the addition
of the sleep handler feature. Similarly, the feature requires
`org.freedesktop.login1` D-Bus interface provided by `systemd-logind`,
so we should allow building without it for systems that do not provide
this (or equivalent) service.