image analyzer: do not toggle sensitivity of windows
Instead of toggling sensitivity of (sub)windows, toggle sensitivity
of their top-level container (grid or box) instead.

If we disable the whole window, we cannot close it anymore, due to
its close button being disabled; not sure if this was always an
issue, or if it happens with contemporary Gtk3 versions...

Toggling the sensitivity on container has the same effect we were
originally aiming for, i.e., enabling/disabling all widgets in the
window.