libMirage, daemon: fix compiler warnings instead of suppressing them
Fix the `-Wcast-function-type` warnings instead of suppressing them.

In daemon, we need to use `G_SOURCE_FUNC()` macro for connecting
`cdemu_device_io_handler` callback (due to number of parameters in the
involved callback type); for the sake of consistency, use the macro
everywhere where cast to `GSourceFunc` is required. Since the macro
became available in GLib >= 2.58, and we still (at least nominally)
support building with older versions (2.38), provide a fallback
definition.

Replace manual casts to `GCallback` in both daemon and libMirage with
`G_CALLBACK()` macro - mostly for the sake of consistency.