libMirage: MDX parser: add support for password-less data encryption
It seems that images made with certain profiles (i.e., TAGES) end
up having encrypted image data, without user actually having to
supply a password.

In this case, the password seems to be derived from the salt data
at the start of encryption header, i.e., same scheme that is used
for encryption of descriptor itself.

So when encountering user data encryption header, first try to
decipher it using password derived from salt data; and only if
that fails, prompt for user password.